+++ src/pages/404.tsx
@@ -1,0 +1,8 @@
+export default function WrongPage() {
+  return (
+    <div>
+      <h1>404</h1>
+      <p>Page not found</p>
+    </div>
+  );
+}
